requirements.txt
xgboost
joblib
flask
flask_restful
flask_jsonpify
flask_cors
python-binance
sklearn
IPython
pymysql
pandas
numpy
matplotlib
ta
tensorflow>=2.0.0a
absl-py